+++ packages/core/src/RenderingEngine/renderPasses/index.ts
@@ -1,0 +1,3 @@
+import { createSharpeningRenderPass } from './sharpeningRenderPass';
+
+export { createSharpeningRenderPass };